requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
django-crum==0.7.3
django-model-utils==3.1.2
django-waffle==0.15.1     # via edx-django-utils, edx-drf-extensions
django==1.11.20
djangorestframework-jwt==1.11.0  # via edx-drf-extensions
djangorestframework==3.9.2  # via edx-drf-extensions, rest-condition
edx-django-utils==1.0.3   # via edx-drf-extensions
edx-drf-extensions==2.0.1
edx-opaque-keys==0.4.4    # via edx-drf-extensions
future==0.17.1            # via pyjwkest
idna==2.8                 # via requests
newrelic==4.14.0.115      # via edx-django-utils
pbr==5.1.3                # via stevedore
psutil==1.2.1             # via edx-django-utils, edx-drf-extensions
pycryptodomex==3.7.3      # via pyjwkest
pyjwkest==1.3.2           # via edx-drf-extensions
pyjwt==1.7.1              # via djangorestframework-jwt
pymongo==3.7.2            # via edx-opaque-keys
python-dateutil==2.8.0    # via edx-drf-extensions
pytz==2018.9              # via django
requests==2.21.0          # via edx-drf-extensions, pyjwkest
rest-condition==1.0.3     # via edx-drf-extensions
semantic-version==2.6.0   # via edx-drf-extensions
six==1.12.0
stevedore==1.30.1         # via edx-opaque-keys
urllib3==1.24.1           # via requests
